+++ packages/core/src/metas/registerInitial.ts
@@ -1,4 +1,5 @@
 import type { Faro } from '../sdk';
+import { getBundleId } from '../utils/sourceMaps';
 import { VERSION } from '../version';
 
 import type { Meta } from './types';
@@ -8,6 +9,9 @@ export function registerInitialMetas(faro: Faro): void {
     sdk: {
       version: VERSION,
     },
+    app: {
+      bundleId: faro.config.app.name && getBundleId(faro.config.app.name),
+    },
   };
 
   const session = faro.config.sessionTracking?.session;
@@ -16,7 +20,7 @@ export function registerInitialMetas(faro: Faro): void {
   }
 
   if (faro.config.app) {
-    initial.app = faro.config.app;
+    initial.app = { ...faro.config.app, ...initial.app };
   }
 
   if (faro.config.user) {