+++ packages/core/src/utils/sourceMaps.ts
@@ -1,0 +1,5 @@
+import { globalObject } from '../globalObject';
+
+export function getBundleId(appName: string): string | undefined {
+  return (globalObject as any)?.[`__faroBundleId_${appName}`];
+}